requirements.txt
alembic==1.7.4
anyio==3.3.4
asgiref==3.4.1
black==21.9b0
click==8.0.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.3.0
Mako==1.1.5
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pydantic==1.8.2
python-dotenv==0.19.1
regex==2021.10.23
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.16.0
tomli==1.2.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
uvloop==0.16.0
zipp==3.6.0
